+++ src/model/Model_Splittransaction.h
@@ -62,7 +62,7 @@ public:
     static double get_total ( const std::vector<Split> &local_splits );
     static const wxString get_tooltip ( const std::vector<Split> &local_splits, const Model_Currency::Data *currency );
     std::map<int, Model_Splittransaction::Data_Set> get_all();
-    int update ( const Data_Set &rows, int transactionID );
+    size_t update ( const Data_Set &rows, int transactionID );
 };
 
 #endif
